requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
psycopg2-binary
python-dotenv
pydantic
streamlit
requests
sqlalchemy
asyncpg
pytest
pytest-asyncio
httpx
pandas
openpyxl
python-docx
requests
